SIMPLY-2536 Fix / silence warnings per Xcode 11.3.1
Rationale behind these fixes vs suppressions:
- turn off warnings that are innocuous (unused functions, unreachable code) at project level, inline otherwise
- fix missed `override` declarations
- fix C++ member initialization order (where compiler would have changed the order anyway, e.g. `cfi.h`)
- fix any warning where a simple/unique solution exists (e.g. `filter.h`, `zip.h`, `zip_ftell.c`))
- fix any warning related to likely bugs (e.g. `zip_dirent.c`, `iri.cpp`)
- suppress warnings by type-casting if non-lossy because of existing in-code checks (e.g. `zip_fseek.c`)
- suppress warnings generating from iOS SDK (e.g. `filter_manager_impl.h`)
- suppress warnings related to implicit casting in calls to iOS SDK functions whose params should likely have different types (e.g. `manifest.cpp`)
- remove some redundant imports proliferating warnings (e.g. `epub_collection.cpp`, `link.cpp`)

+++ ePub3/ePub/manifest.cpp
@@ -272,9 +272,9 @@ shared_ptr<xml::Document> ManifestItem::ReferencedDocument() const
 	
 	xmlDocPtr raw;
     if ( _mediaType == "text/html" ) {
-        raw = htmlReadMemory((const char*)docBuf, resbuflen, path.c_str(), encoding, flags);
+        raw = htmlReadMemory((const char*)docBuf, (int)resbuflen, path.c_str(), encoding, flags);
     } else {
-        raw = xmlReadMemory((const char*)docBuf, resbuflen, path.c_str(), encoding, flags);
+        raw = xmlReadMemory((const char*)docBuf, (int)resbuflen, path.c_str(), encoding, flags);
     }
 	
 	result = xml::Wrapped<xml::Document>(raw);
